+++ Core/DbContext.cs
@@ -319,7 +319,7 @@ namespace Greatbone.Core
         // RESULTSET
         //
 
-        public D ToObject<D>(ushort proj = 0x00ff) where D : IData, new()
+        public D ToObject<D>(int proj = 0x00ff) where D : IData, new()
         {
             D obj = new D();
             obj.Read(this, proj);
@@ -334,7 +334,7 @@ namespace Greatbone.Core
             return obj;
         }
 
-        public D[] ToArray<D>(ushort proj = 0x00ff) where D : IData, new()
+        public D[] ToArray<D>(int proj = 0x00ff) where D : IData, new()
         {
             List<D> coll = new List<D>(32);
             while (Next())
@@ -354,7 +354,7 @@ namespace Greatbone.Core
             return coll.ToArray();
         }
 
-        public List<D> ToList<D>(ushort proj = 0x00ff) where D : IData, new()
+        public List<D> ToList<D>(int proj = 0x00ff) where D : IData, new()
         {
             List<D> coll = new List<D>(32);
             while (Next())
@@ -374,7 +374,7 @@ namespace Greatbone.Core
             return coll;
         }
 
-        public Dictionary<K, D> ToDictionary<K, D>(Func<D, K> keyer, ushort proj = 0x00ff) where D : IData, new()
+        public Dictionary<K, D> ToDictionary<K, D>(Func<D, K> keyer, int proj = 0x00ff) where D : IData, new()
         {
             int initial = doerctx.Doer?.Limit ?? 0;
             if (initial <= 0) initial = 32;
@@ -587,7 +587,7 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public bool Get<D>(string name, ref D v, ushort proj = 0x00ff) where D : IData, new()
+        public bool Get<D>(string name, ref D v, int proj = 0x00ff) where D : IData, new()
         {
             try
             {
@@ -726,7 +726,7 @@ namespace Greatbone.Core
             return false;
         }
 
-        public bool Get<D>(string name, ref D[] v, ushort proj = 0x00ff) where D : IData, new()
+        public bool Get<D>(string name, ref D[] v, int proj = 0x00ff) where D : IData, new()
         {
             try
             {
@@ -960,7 +960,7 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public IDataInput Let<D>(out D v, ushort proj = 0x00ff) where D : IData, new()
+        public IDataInput Let<D>(out D v, int proj = 0x00ff) where D : IData, new()
         {
             try
             {
@@ -989,7 +989,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public IDataInput Let<D>(out D[] v, ushort proj = 0x00ff) where D : IData, new()
+        public IDataInput Let<D>(out D[] v, int proj = 0x00ff) where D : IData, new()
         {
             try
             {
@@ -1038,14 +1038,14 @@ namespace Greatbone.Core
             BufferUtility.Return(dcont); // back to pool
         }
 
-        public void Publish(string name, string shard, int arg, IData obj, ushort proj = 0x00ff)
+        public void Publish(string name, string shard, int arg, IData obj, int proj = 0x00ff)
         {
             JsonContent cont = new JsonContent(true).Put(null, obj, proj);
             Publish(name, shard, arg, cont);
             BufferUtility.Return(cont); // back to pool
         }
 
-        public void Publish<D>(string name, string shard, int arg, D[] arr, ushort proj = 0x00ff) where D : IData
+        public void Publish<D>(string name, string shard, int arg, D[] arr, int proj = 0x00ff) where D : IData
         {
             JsonContent cont = new JsonContent(true).Put(null, arr, proj);
             Publish(name, shard, arg, cont);
